perf(bmcu): stop paying 3848 bytes of flash for four doubles
Four lines in the pull calibration were the only doubles in the firmware:
two accumulators summing 90 float ADC samples, and the two divisions that
averaged them. They dragged in the entire libgcc soft-double family --
__adddf3, __divdf3, __truncdfsf2, __extendsfdf2 -- on a part with 61440
bytes of flash and under 4 KB spare.

Measured on the worst-case matrix variant (soft_load, autoload, rgb, AMS_B,
0.90f), identical config either side of the change:

  57676 -> 53828 bytes, 3848 saved
  headroom 3764 -> 7612, so it more than doubles

No double helper symbol remains linked. RAM drops 4 bytes.

Precision is not the constraint. Simulated over 20000 randomised runs of 90
samples across 0.05-3.30 V, worst accumulated error is 3.8e-4 V in the sum
and 4.2e-6 V in the mean -- half a percent of one 12-bit ADC LSB, and far
under the ADC noise the 90-sample average exists to suppress.

It is still a behavioural change, narrowly: these means become
MC_PULL_V_OFFSET and MC_DM_KEY_NONE_THRESH in the calibration sector. It
runs at first boot only, never on the printer bus path, and stored
calibration on already-provisioned boards is untouched -- only a board that
recalibrates after this lands sees the difference, which is below what the
ADC can resolve.

Found by a flash-footprint investigation that also confirmed every standard
size flag is already enabled, so this was the one large win available.

+++ src/MC_PULL_calibration.cpp
@@ -337,8 +337,19 @@ void MC_PULL_calibration_boot()
 
     const bool ok_wipe = Flash_NVM_full_clear();
 
-    double sum_raw[4] = {0, 0, 0, 0};
-    double sum_key[4] = {0, 0, 0, 0};
+    // float, not double, and it matters far more than it looks. These four
+    // lines were the only doubles in the firmware, and they dragged in the
+    // whole libgcc soft-double family -- __adddf3, __divdf3, __truncdfsf2,
+    // __extendsfdf2, 3752 bytes measured -- on a part with 61440 bytes of
+    // flash and under 4 KB spare. Nothing else here needs the range.
+    //
+    // Precision is not the constraint. Simulated over 20000 randomised runs
+    // of 90 samples spanning the full 0.05-3.30 V range, worst accumulated
+    // error is 3.8e-4 V in the sum and 4.2e-6 V in the mean -- half a percent
+    // of one 12-bit ADC LSB (8.1e-4 V), and far below the noise these 90
+    // samples exist to average out in the first place.
+    float sum_raw[4] = {0, 0, 0, 0};
+    float sum_key[4] = {0, 0, 0, 0};
     const int N = 90;
     const uint32_t tpm = time_hw_ticks_per_ms();
     const uint32_t t0  = time_ticks32();
@@ -367,7 +378,7 @@ void MC_PULL_calibration_boot()
     for (int ch = 0; ch < 4; ch++)
     {
         if (!filament_channel_inserted[ch]) continue;
-        center_raw[ch] = (float)(sum_raw[ch] / (double)N);
+        center_raw[ch] = sum_raw[ch] / (float)N;
     }
 
     for (int ch = 0; ch < 4; ch++)
@@ -383,7 +394,7 @@ void MC_PULL_calibration_boot()
         MC_PULL_V_OFFSET[ch] = 1.65f - center_raw[ch];
         MC_PULL_POLARITY[ch] = 1;
         MC_DM_KEY_NONE_THRESH[ch] =
-            dm_key_none_threshold_from_idle((float)(sum_key[ch] / (double)N));
+            dm_key_none_threshold_from_idle(sum_key[ch] / (float)N);
     }
 
     float center_v_ref[4] = {1.65f, 1.65f, 1.65f, 1.65f};
